+++ Archive/Repcount and Flex Sensor Prototype/Posture Detection System Artifact.pdf
