+++ Archive/Repcount and Flex Sensor Prototype/Rep Count System Artifact.pdf
