+++ Archive/Research/Cycling/Research into sensors (for cycling prototype).docx
