+++ Archive/Research/Smart Watches/Capabilities of Fitbit for measuring exercise effectiveness.docx
